significant refactor to prepare for multiple backends
Previously, CLI functions were directly dependant on the remote backend
for storing data. This made it difficult to implement another local
backend without breaking the design.

The backend store should now implement a trait which is used by the CLI,
details of backend implementation is preferebly contained to the
implementation of that trait, rather than leaking to CLI.