Introduce context concept, defer current impl to rpw remote
Introducing concept makes the CLI a bit more clunky to use. But has the
added benefit of making the interface easier to understand.

And let's face it, much more importantly, makes it easier to implement
to backends.

For later consideration is whether the user could  switch between
remote/local backend without providing the context string each time.

Buuuut that is for later consideration.

Signed-off-by: Patrik Lundgren <patrik.lundgren@outlook.com>