Implement FromStr for vault so that arg code can be used 'directly'
Not sure if this is a good design, but it enables the use of the command
error code to catch all errors from parses to conversion to file system
reads.

It simplifies the code and can center the functionality on the cli.
Whilst not requiring intermediary APIs which perform the same
conversions.

We'll see how it fares :D

Signed-off-by: Patrik Lundgren <patrik.lundgren@outlook.com>