Implement export/import to/from json plain-text format
Add commands import|export file <FNAME> which take a json file:

{ [ { "id": "foo", "pw": "bar" }, { "id" : "lorem", "pw": "ipsum"}] }

This format (and functions) hopefully makes it easier to
import passwords from other password mangers.

In addition to this, import/export to plaintext format may be useful
when migrating the vault format. Which is likelier (famous last words)
to change.

Signed-off-by: Patrik Lundgren <patrik.lundgren@outlook.com>