Fetch vault and master password via program state
Previously, a flag was appended to all commands (-v <vaultname>)
and a flag (--password <password). This doesn't play nice with
the design of 'clap' as commands which did not require these
flags complained 'unexpected option' or some such.

Instead store vault in program state and use 'chain-of-command'
type getter.

state.vault ?: --v <vault> ?: prompt ?: fail

There is one caveat though. We may not want to prompt but to
require the 'vault' option to be passed by flag option or
by state. This case is not supported and will fail on an
unwrap. Clap handled this before.

Signed-off-by: Patrik Lundgren <patrik.lundgren@outlook.com>